Change SonataFlow services address
This PR is designed to enable a smoother experience with working against
a local workflow runtime.

The reason for replacing 'localhost' with a 'sonataflow' is due to the fact
that some node versions fail to resolve the address properly. Hence, using
`http://sonataflow:8080` and adding 'sonataflow' to /etc/hosts to be resolved
as 127.0.0.1 will allow running all on the same machine for development purposes.

This decision will be revisited when the plugin is updated to work with production backing services.

Signed-off-by: Moti Asayag <masayag@redhat.com>